Cleanup some applause code
Cleans some applause code
