+++ client/src/app/shared/components/applause-bar-display/applause-bar-display.component.html
