+++ client/src/app/shared/components/applause-particle-display/applause-particle-display.component.html
@@ -1,0 +1,3 @@
+<div [osResized]="resizeSubject" class="particle-wrapper">
+    <Particles id="particles" [options]="options" (particlesLoaded)="particlesLoaded($event)"> </Particles>
+</div>
