+++ client/src/app/shared/components/applause-particle-display/applause-particle-display.component.scss
