--- client/src/app/shared/components/particle-display/particle-display.component.html
@@ -1,3 +1,0 @@
-<div [osResized]="resizeSubject" class="particle-wrapper">
-    <Particles id="particles" [options]="particlesOptions" (particlesLoaded)="particlesLoaded($event)"> </Particles>
-</div>
