+++ client/src/app/shared/components/vjs-player/vjs-player.component.html
@@ -1,5 +1,5 @@
 <div class="video-wrapper">
-    <os-particle-display *ngIf="showParticles" class="applause-particles"></os-particle-display>
+    <os-applause-particle-display *ngIf="showParticles" class="applause-particles"></os-applause-particle-display>
     <div class="player-container" [ngClass]="{ hide: !isUrlOnline }">
         <video #videoPlayer class="video-js" controls preload="none"></video>
     </div>
